Restore appearance of upscaled tiles
Explicitly request the upscaling filter we want, to ensure compatibility with
different Pillow versions.

In the qud-wiki project, upgrading Pillow to 7.0.0 caused the upscaling of our
big tiles to be done by a bicubic filter instead of nearest pixel. This caused
blurring in the big image.

For more information on the Pillow change:
https://pillow.readthedocs.io/en/stable/releasenotes/7.0.0.html#default-resampling-filter