Completely overhaul types of QudObject properties
When this project was first started, QudObject was a backing class for
the wikification process, and a lot of the wiki logic was baked into
it (templates, etc.) Although this was mostly split off into qud-wiki,
a lot of the properties we define were still returning strings from XML
for historical reasons.

This commit changes all the properties to return Pythonic types and
values such as True/False, integers, and lists instead of 'yes'/'no',
integers as strings, and comma-separated lists. This was done because
Hagadias is now becoming the data source for more projects.

A separate commit will be coming to restore the proper formatting in the
qud-wiki project.