Fix repeatedly checking if object has tile
The _tile attribute of QudObjects was being created in __init__ as
None, but then later if _tile is None we were checking every time to
see if a the object is eligible to have a tile.

Change this to only initialize the _tile attribute to None if the
QudObject is not eligible for a tile, and otherwise do not have this
cache attribute at all.