two-pass load for ObjectBlueprints.xml
Make sure that we're loading the full XML before we try to resolve any object inheritance. This more closely mimics how the game loads ObjectBlueprints.xml, and makes QBE resiliant against scenarios where an object inherits from a different object that is not defined until later in the XML file.